+++ NEWS.md
@@ -1,3 +1,6 @@
+# rpcdas 0.1.1
+* Adds get_sinasc function with basic filters.
+
 # rpcdas 0.1.0
 
-* First version with initials funcions.
+* First version with get_sim and other basic functions.